// src/app/Models/models.ts
export interface Education {
    institute : string;
    course : string;
    duration : string;
    
}

export interface WorkExperience{
    role: string;
    company: string;
    duration: string;
    description: string[] ;
}
export interface Skill {
    name: string;
    level: string;
    rating : number;
}

export interface Projects{
    title: string;
    technologies: string;
    description: string[]
}